Add files via upload
This Time-based prospective memory task requires participants to press the 'y' every time a minute has elapsed. They can monitor they time elapsed by pressing the space bar, which will make a small digital clock appear for about 1 and a half second. The whole task lasts about 4 minutes and a half. 113 words and 113 non-words are displayed on the screen for 0.5 s, separated by a fixation cross which will be on the screen for 0.725 s. The "yes" and "no" button are the "m" and "c" keys of the computer, respectively. 
Combine the lexical decision task practice trial with this one to have a coplete prospective memory task.